+++ waffle-chai/.mocharc.js
@@ -2,5 +2,6 @@ process.env.NODE_ENV = 'test'
 module.exports = {
   require: 'ts-node/register/transpile-only',
   timeout: 50000,
-  spec: 'test/**/*.{js,ts}'
+  spec: 'test/**/*.{js,ts}',
+  file: 'test/test-setup.ts'
 }