+++ waffle-chai/src/matchers/revertedWith.ts
@@ -87,11 +87,11 @@ const decodeHardhatError = (error: any, context: any) => {
       }
     }
     {
-      const regexp = new RegExp('reverted with reason string "(.*?)"');
+      const regexp = new RegExp('revert(ed)? with reason (string )?"(.*?)"');
 
       const matches = regexp.exec(errorString);
       if (matches && matches.length >= 1) {
-        return matches[1];
+        return matches[matches.length - 1];
       }
     }
     return undefined;